requirements.txt
alembic==1.14.0
asyncpg==0.30.0
bcrypt==4.2.1
celery==5.4.0
fastapi==0.115.6
fastapi-cache2==0.2.2
httpx==0.28.1
Jinja2==3.1.4
passlib==1.7.4
pydantic==2.10.4
pydantic-settings==2.7.0
pytest==8.3.4
pytest-asyncio==0.25.0
python-jose==3.3.0
redis==5.2.1
sqladmin==0.20.1
SQLAlchemy==2.0.36
uvicorn==0.34.0